fixed the test and added the flag to other options (#1)
* feat: Update macOS options to include hidden and callback functionality

* feat: updated test for export cli required macos
